backend/requirements.txt
fastapi
pydantic
uvicorn
pymongo
sentence_transformers
python-dotenv
